+++ src/main/java/com/cheolhyeon/diary/app/event/s3/S3EventListener.java
@@ -1,4 +1,4 @@
-package com.cheolhyeon.diary.diary.dto;
+package com.cheolhyeon.diary.app.event.s3;
 
 
 import com.cheolhyeon.diary.diary.service.S3Service;
